--- LLM/Mistral/getAPIkey.py
@@ -1,8 +1,0 @@
-import os
-
-
-def make_api_key():
-    """
-    place your API key in environment
-    """
-    os.environ["MISTRAL_API_KEY"] = ВАШ API КЛЮЧ
