+++ LLM/app/utils/Mistral/mistral.py
@@ -1,12 +1,16 @@
 from mistralai import Mistral
 import os
+
+from ...utils.Mistral.config import api_key
+
+
 # from getAPIkey import make_api_key
 
 def make_api_key():
     """
     place your API key in environment
     """
-    os.environ["MISTRAL_API_KEY"] =
+    os.environ["MISTRAL_API_KEY"] = api_key
 
 def initialize_client():
     """Инициализирует клиента Mistral с API-ключом."""
